Pangram
Input Format
Input consists of a string s.

Constraints 
Length of  can be at most 10Pow3(1<=|s|<=10Pow3)and it may contain spaces, lower case and upper case letters. Lower-case and upper-case instances of a letter are considered the same.

Output Format
Output a line containing pangram if s is a pangram, otherwise output not pangram.