Two String
Input Format
Several test cases will be given to you in a single file. The first line of the input will contain a single integer , the number of test cases.

Then there will be T descriptions of the test cases. Each description contains two lines. The first line contains the string A and the second line contains the string B.

Output Format
For each test case, display YES (in a newline), if there is a common substring. Otherwise, display NO.